requirements.txt
regex>=2024.4.16
symspellpy>=6.7.7
transformers>=4.43.0
torch>=2.2.0
symspellpy>=6.7.7
pyspellchecker>=0.8.1

